+++ packages/nextjs/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @savvagent/nextjs
 
+## 1.0.1
+
+### Patch Changes
+
+- Updates for new API parameters
+- Updated dependencies
+  - @savvagent/react@1.0.1
+  - @savvagent/sdk@1.0.1
+
 ## 1.0.0
 
 ### Major Changes