+++ packages/node-server/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @savvagent/node-server
 
+## 1.0.1
+
+### Patch Changes
+
+- Updates for new API parameters
+
 ## 1.0.0
 
 ### Major Changes