+++ packages/solid/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @savvagent/solid
 
+## 1.0.1
+
+### Patch Changes
+
+- Updates for new API parameters
+- Updated dependencies
+  - @savvagent/sdk@1.0.1
+
 ## 1.0.0
 
 ### Major Changes